requirements.txt
torch==2.1.0
librosa==0.10.1
openai-whisper
soundfile==0.12.1
pathlib~=1.0.1